Move all icon-related code to horizons.world.icons
