+++ horizons/world/icons/__init__.py
@@ -18,3 +18,6 @@
 # Free Software Foundation, Inc.,
 # 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 # ###################################################
+
+from production_finished import ProductionFinishedIcon
+from problem_status import *
