+++ horizons/world/icons/abstracticonmanager.py
