+++ horizons/world/icons/production_finished.py
