+++ horizons/world/icons/productionfinishediconmanager.py
@@ -26,8 +26,8 @@ from horizons.constants import GAME_SPEED
 from horizons.messaging import ResourceProduced
 from horizons.scheduler import Scheduler
 from horizons.util.python.callback import Callback
-from horizons.world.managers.abstracticonmanager import AbstractIconManager
-from horizons.world.production_finished import ProductionFinishedIcon
+from horizons.world.icons.abstracticonmanager import AbstractIconManager
+from horizons.world.icons import ProductionFinishedIcon
 
 class ProductionFinishedIconManager(AbstractIconManager):
 	"""Manager class that manages all production finished icons. It listens to
